Cambio del tipo de dato int a byte en option  para ahorrar espacio en memoria

+++ divideYVenceras/Program.cs
@@ -80,7 +80,7 @@ Console.WriteLine();
 Console.WriteLine("4.Salir del Menu.");
 
 bool bandera = true;
-int option = 0;
+byte option = 0;
 
 while (bandera)
 {
